Module Hegel_ffi.Loader

val version : string

The libhegel version these bindings target.

val checksums : (string * string) list

val release_base : string

val getenv_nonempty : string -> string option

val read_line_cmd : string -> string

val os_id : unit -> string

val arch_id : unit -> string

val ext_of_os : string -> string

val local_basename : string -> string

val release_artifact : string -> string -> string

val is_dir : string -> bool

val is_file : string -> bool

val home : unit -> string

val cache_dir : unit -> string

val mkdir_p : Stdlib.String.t -> unit

val sha256_of_file : string -> string

val from_env : string -> string option

val from_site : unit -> string option

val from_sibling : string -> string option

val download_verified : url:string -> expected:Stdlib.String.t -> cache_path:string -> string

val from_cache_or_download : string -> string -> string

val locate : unit -> string

locate () returns the path to a usable libhegel shared library, downloading and caching it if necessary. Raises Failure with a descriptive message if no library can be found or fetched.